Moved components into pages folder
