+++ frontend/src/components/pages/MyGamesLibrary.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState, useContext } from "react";
-import GameCard from "./GameCard";
-import { getGamesByUsername } from "../utils/sanity/userServices";
-import UserContext from "../context/UserContext";
+import GameCard from "./../GameCard";
+import { getGamesByUsername } from "../../utils/sanity/userServices";
+import UserContext from "../../context/UserContext";
 
 export default function MyGamesLibrary() {
   const { user } = useContext(UserContext);
